+++ rust/polars-ocaml/src/misc.rs
@@ -1,5 +1,6 @@
-use crate::utils::PolarsDataType;
-use crate::utils::*;
+use crate::interop::*;
+use crate::polars_types::PolarsDataType;
+use crate::polars_types::*;
 use ocaml_interop::{DynBox, OCaml, OCamlList, OCamlRef, ToOCaml};
 use polars::prelude::*;
 use polars_ocaml_macros::ocaml_interop_export;
