+++ rust/polars-ocaml/src/series.rs
@@ -1,4 +1,5 @@
-use crate::utils::*;
+use crate::interop::*;
+use crate::polars_types::*;
 use chrono::naive::{NaiveDate, NaiveDateTime, NaiveTime};
 use chrono::Duration;
 use ocaml_interop::{
